requirements.txt
streamlit
langchain
langgraph
langchain-openai
langchain-community
faiss-cpu
rank-bm25
rapidfuzz
tiktoken
python-dotenv
beautifulsoup4
requests
nltk
langsmith
fastapi
uvicorn[standard]
python-dotenv
huggingface_hub
typing_extensions
# Evaluation dependencies
pandas
matplotlib
seaborn